Fix kappa_calculation import loop
