+++ tardis/energy_input/__init__.py
@@ -1,5 +1,3 @@
 """
 Contains classes and functions to handle energy deposition and transport.
 """
-
-from tardis.energy_input.util import *
